attn_res_weighted_reduce_nc: move the op under deepseek_prefill
The prefill team builds and maintains it, and that domain registers one CMake
target per op instead of sharing a group library. The Python name is unchanged
— bind_function sets it, not the namespace.
