+++ ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/CMakeLists.txt
@@ -1,5 +1,6 @@
 # Owned by @tenstorrent/metalium-developers-ds-prefill. Add/remove ops for
 # this domain here without touching ttnn/CMakeLists.txt.
+add_subdirectory(attn_res_weighted_reduce_nc)
 add_subdirectory(combine)
 add_subdirectory(dispatch)
 add_subdirectory(extract)
@@ -27,6 +28,7 @@ add_library(TTNN::Ops::Experimental::DeepSeekPrefill ALIAS ttnn_op_deepseek_pref
 target_link_libraries(
     ttnn_op_deepseek_prefill
     INTERFACE
+        TTNN::Ops::Experimental::DeepSeekPrefill::AttnResWeightedReduceNc
         TTNN::Ops::Experimental::DeepSeekPrefill::Combine
         TTNN::Ops::Experimental::DeepSeekPrefill::Dispatch
         TTNN::Ops::Experimental::DeepSeekPrefill::Extract
@@ -61,6 +63,7 @@ target_link_libraries(
 target_sources(
     ttnn_op_deepseek_prefill
     INTERFACE
+        $<TARGET_OBJECTS:TTNN::Ops::Experimental::DeepSeekPrefill::AttnResWeightedReduceNc>
         $<TARGET_OBJECTS:TTNN::Ops::Experimental::DeepSeekPrefill::Combine>
         $<TARGET_OBJECTS:TTNN::Ops::Experimental::DeepSeekPrefill::Dispatch>
         $<TARGET_OBJECTS:TTNN::Ops::Experimental::DeepSeekPrefill::Extract>
