+++ ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/attn_res_weighted_reduce_nc/attn_res_weighted_reduce_nc_nanobind.hpp
@@ -6,9 +6,9 @@
 
 #include "ttnn-nanobind/nanobind_fwd.hpp"
 
-namespace ttnn::operations::experimental::reduction::detail {
+namespace ttnn::operations::experimental::deepseek_prefill::attn_res_weighted_reduce_nc::detail {
 
 namespace nb = nanobind;
 
 void bind_attn_res_weighted_reduce_nc(nb::module_& mod);
-}  // namespace ttnn::operations::experimental::reduction::detail
+}  // namespace ttnn::operations::experimental::deepseek_prefill::attn_res_weighted_reduce_nc::detail
