+++ ttnn/cpp/ttnn/operations/experimental/deepseek_prefill/attn_res_weighted_reduce_nc/device/kernels/reader_weighted_reduce_nc.cpp
