+++ ttnn/cpp/ttnn/operations/experimental/reduction/sources.cmake
@@ -5,7 +5,6 @@ set(TTNN_OP_EXPERIMENTAL_REDUCTION_API_HEADERS
     deepseek_grouped_gate/deepseek_grouped_gate.hpp
     deepseek_moe_fast_reduce_nc/deepseek_moe_fast_reduce_nc.hpp
     fast_reduce_nc/fast_reduce_nc.hpp
-    attn_res_weighted_reduce_nc/attn_res_weighted_reduce_nc.hpp
     integral_image/intimg.hpp
 )
 
@@ -13,9 +12,6 @@ set(TTNN_OP_EXPERIMENTAL_REDUCTION_SRCS
     fast_reduce_nc/device/fast_reduce_nc_device_operation.cpp
     fast_reduce_nc/device/fast_reduce_nc_program_factory.cpp
     fast_reduce_nc/fast_reduce_nc.cpp
-    attn_res_weighted_reduce_nc/device/attn_res_weighted_reduce_nc_device_operation.cpp
-    attn_res_weighted_reduce_nc/device/attn_res_weighted_reduce_nc_program_factory.cpp
-    attn_res_weighted_reduce_nc/attn_res_weighted_reduce_nc.cpp
     deepseek_moe_fast_reduce_nc/device/deepseek_moe_fast_reduce_nc_device_operation.cpp
     deepseek_moe_fast_reduce_nc/device/deepseek_moe_fast_reduce_nc_program_factory.cpp
     deepseek_moe_fast_reduce_nc/deepseek_moe_fast_reduce_nc.cpp
@@ -38,7 +34,6 @@ set(TTNN_OP_EXPERIMENTAL_REDUCTION_SRCS
 set(TTNN_OP_EXPERIMENTAL_REDUCTION_NANOBIND_SRCS
     fast_reduce_nc/fast_reduce_nc_nanobind.cpp
     fast_reduce_nc/fast_reduce_nc_nanobind.cpp
-    attn_res_weighted_reduce_nc/attn_res_weighted_reduce_nc_nanobind.cpp
     deepseek_moe_fast_reduce_nc/deepseek_moe_fast_reduce_nc_nanobind.cpp
     deepseek_moe_fast_reduce_nc_fused/deepseek_moe_fast_reduce_nc_fused_nanobind.cpp
     integral_image/intimg_nanobind.cpp
